# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "BODHI CLINIC SERVICE",
    "description": "This service is used to manage Clinic & Doctor Details of BODHI Application ",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header"
    }
  },
  "security": [
    {
      "JWT": []
    }
  ],
  "responses": {
    "UnauthorizedError": {
      "description": "Access token is missing or invalid"
    }
  },
  "host": "localhost:3002",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "Clinic",
      "description": "API for Clinic details in the system"
    },
    {
      "name": "Doctor",
      "description": "API for Doctor details in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/clinic": {
      "get": {
        "tags": ["Clinic"],
        "summary": "Get all clinics in system",
        "description": "This can only be done by the logged in user.",
        "operationId": "viewClinics",
        "produces": "application/xml,application/json",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Clinic"
            }
          }
        }
      },
      "post": {
        "tags": ["Clinic"],
        "summary": "Insert clinic details into system",
        "description": "This can only be done by the logged in user.",
        "operationId": "insertClinic",
        "produces": "application/xml,application/json",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": "true",
            "description": "Enter clinic details"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Clinic"
            }
          }
        }
      },
      "put": {
        "tags": ["Clinic"],
        "summary": "Update clinic details",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateClinic",
        "produces": "application/xml,application/json",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": "true",
            "description": "Update clinic"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Clinic"
            }
          }
        }
      }
    },
    "/clinic/name/{name}": {
      "get": {
        "tags": ["Clinic"],
        "summary": "View Clinic details by name",
        "description": "This can only be done by the logged in user.",
        "operationId": "viewClinicByName",
        "produces": "application/xml,application/json",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": "true",
            "description": "View Clinic Details"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Clinic"
            }
          }
        }
      }
    },
    "/clinic/loadCache/": {
      "get": {
        "tags": ["Clinic"],
        "summary": "Load Cache",
        "description": "This can only be done by the logged in user.",
        "operationId": "loadCache",
        "produces": "application/xml,application/json",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Clinic"
            }
          }
        }
      }
    },
    "/clinic/readFromCache/{clinic_name}": {
      "get": {
        "tags": ["Clinic"],
        "summary": "Read from Cache",
        "description": "This can only be done by the logged in user.",
        "operationId": "readFromCache",
        "produces": "application/xml,application/json",
        "parameters": [
          {
            "name": "clinic_name",
            "in": "path",
            "required": "true",
            "description": "Enter Clinic Name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Clinic"
            }
          }
        }
      }
    },
    "/clinic/schedule/{clinic_id}/{doctor_id}": {
      "get": {
        "tags": ["Clinic"],
        "summary": "Get Schedule by Doctor ID and Date",
        "description": "This can only be done by the logged in user.",
        "operationId": "getScheduleByIdAndDate",
        "produces": "application/xml,application/json",
        "parameters": [
          {
            "name": "clinic_id",
            "in": "path",
            "required": "true",
            "description": "Enter Clinic ID"
          },
          {
            "name": "doctor_id",
            "in": "path",
            "required": "true",
            "description": "Enter Doctor ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Clinic"
            }
          }
        }
      }
    },
    "/clinic/id/{id}": {
      "get": {
        "tags": ["Clinic"],
        "summary": "View Clinic details by ID",
        "description": "This can only be done by the logged in user.",
        "operationId": "viewClinicByID",
        "produces": "application/xml,application/json",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": "true",
            "description": "View Clinic Details"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Clinic"
            }
          }
        }
      }
    },
    "/clinic/city/{city}": {
      "get": {
        "tags": ["Clinic"],
        "summary": "View Clinic details by City Name",
        "description": "This can only be done by the logged in user.",
        "operationId": "viewClinicBycity",
        "produces": "application/xml,application/json",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": "true",
            "description": "View Clinic Details"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Clinic"
            }
          }
        }
      }
    },
    "clinic/doctor/{id}": {
      "get": {
        "tags": ["Doctor"],
        "summary": "View Doctor Details",
        "description": "This can only be done by the logged in user.",
        "operationId": "viewDoctors",
        "produces": "application/xml,application/json",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": "true",
            "description": "View Clinic Details"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Doctor"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Clinic": {
      "required": ["name"],
      "properties": {
        "name": {
          "type": "string",
          "uniqueItems": true
        },
        "address": {
          "type": "string"
        },
        "contact_no": {
          "type": "string"
        },
        "about": {
          "type": "string"
        }
      }
    },
    "Doctor": {
      "required": ["first_name", "last_name"],
      "properties": {
        "first_name": {
          "type": "string",
          "uniqueItems": true
        },
        "last_name": {
          "type": "string",
          "uniqueItems": true
        },
        "address": {
          "type": "string"
        },
        "contact_no": {
          "type": "string"
        },
        "about": {
          "type": "string"
        }
      }
    }
  }
}